+++ midvex_customs_op/models/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+
+from . import customs_stage
+from . import customs_document_type
+from . import customs_operation
+from . import customs_document_requirement